+++ NewsGatherer.py
@@ -8,13 +8,6 @@ To Do:
 2. Check to see if I have already tweeted this article
 3. Perhaps tweet out an image, if Image Url is not null?'''
 
-def goo_shorten_url(url): #Used to get shorter url to tweet
-    post_url = 'https://www.googleapis.com/urlshortener/v1/url?key=AIzaSyB7DHtzdPLF9Pl90XFzkw_YD_sr8PI58gA'
-    data = {'longUrl': url}
-    headers = {'content-type': 'application/json'}
-    r = requests.post(post_url, data=json.dumps(data), headers=headers)
-    return r.json()['id']
-
 def get_news_stories(): #TODO change url query to "tall%20ships"
     url = 'https://newsapi.org/v2/everything'
     data = {
@@ -33,7 +26,12 @@ def parse_news_stories(response, story):
         data = { 'title': news['articles'][story]['title'], 'longUrl': news['articles'][story]['url'], 'description': news['articles'][story]['description']}
         return data
 
-'''
-response = get_news_stories()
-story = parse_news_stories(response, 1)
-'''
+def parse_news_title(story):
+    newsTitle = story['title']
+    return newsTitle
+
+#  Testing
+# response = get_news_stories()
+# story = parse_news_stories(response, 1)
+# print story
+# print parse_news_title(story)
